requirements.txt
certifi==2021.5.30
pyspellchecker==0.6.2
